refactor(test): move i18n/test-helpers.js to TypeScript
